Crate jetstream_rpc

JetStream Rpc

Defines Rpc primitives for JetStream. Of note is the Protocol trait which is meant to be used with the service attribute macro.

Re-exports

pub use framer::*;

Modules

client

context

framer

server

Structs

Framed

A unified Stream and Sink interface to an underlying I/O object, using the Encoder and Decoder traits to encode and decode frames.

Tag

Enums

Error

Traits

Decoder

Decoding of frames via buffers.

Encoder

Trait of helper objects to write out messages as bytes, for use with FramedWrite.

Message

A trait representing a message that can be encoded and decoded.

Protocol

Defines the request and response types for the JetStream protocol.