Crate jetstream_rpc

JetStream Rpc

Defines Rpc primitives for JetStream. Of note is the Protocol trait which is meant to be used with the service attribute macro.

Re-exports

pub use framer::*;

Modules

client

context

framer

server

Structs

Framed

A unified Stream and Sink interface to an underlying I/O object, using the Encoder and Decoder traits to encode and decode frames.

Mux

Client Mux

Rerror

Rlerror

RpcCall

Rversion

version – negotiate protocol version

TagPool

Tversion

version – negotiate protocol version

Enums

Encoding

EncodingError

ErrorFrame

Version

VersionFrame

Constants

HEADER_KEY_JETSTREAM_PROTO

MIMETYPE_JETSTREAM

MIMETYPE_JSON

MIMETYPE_OCTET_STREAM

MIMETYPE_XML

RJETSTREAMERROR

RLERROR

RVERSION

TVERSION

Traits

AnyServer

Decoder

Decoding of frames via buffers.

Encoder

Trait of helper objects to write out messages as bytes, for use with FramedWrite.

IntoError

IntoError trait allows conversion of various error types into Jetstream Error.

Message

A trait representing a message that can be encoded and decoded.

Protocol

Defines the request and response types for the JetStream protocol.

Functions

error_to_rmessage

Type Aliases

Error

InFlight

RxStream

TxSink